+++ tests/test_minimal/test_tools/dandi_transfer_tools.py
@@ -6,20 +6,15 @@ from platform import python_version as get_python_version
 from shutil import rmtree
 from tempfile import mkdtemp
 
-import dandi.dandiapi
 import pytest
 from hdmf.testing import TestCase
 from pynwb import NWBHDF5IO
 
-from neuroconv import run_conversion_from_yaml
 from neuroconv.tools.data_transfers import automatic_dandi_upload
 from neuroconv.tools.nwb_helpers import (
     get_default_nwbfile_metadata,
     make_nwbfile_from_metadata,
 )
-
-from ...test_on_data.setup_paths import ECEPHY_DATA_PATH as DATA_PATH
-from ...test_on_data.setup_paths import OUTPUT_PATH
 
 DANDI_API_KEY = os.getenv("DANDI_API_KEY")
 HAVE_DANDI_KEY = DANDI_API_KEY is not None and DANDI_API_KEY != ""  # can be "" from external forks
@@ -106,34 +101,3 @@ class TestAutomaticDANDIUploadNonParallelNonThreaded(TestCase):
             number_of_jobs=1,
             number_of_threads=1,
         )
-
-
-def test_run_conversion_from_yaml_with_dandi_upload():
-    data_tests_base_folder_path = Path(__file__).parent.parent.parent / "test_on_data"
-    path_to_test_yml_files = data_tests_base_folder_path / "test_yaml" / "conversion_specifications"
-    yaml_file_path = path_to_test_yml_files / "GIN_conversion_specification_dandi_upload.yml"
-    run_conversion_from_yaml(
-        specification_file_path=yaml_file_path,
-        data_folder_path=DATA_PATH,
-        output_folder_path=OUTPUT_PATH,
-        overwrite=True,
-    )
-
-    time.sleep(20)  # Give some buffer room for server to process before making assertions against DANDI API
-
-    client = dandi.dandiapi.DandiAPIClient(api_url="https://api-staging.dandiarchive.org/api")
-    dandiset = client.get_dandiset("200560")
-
-    expected_asset_paths = [
-        "sub-yaml-1/sub-yaml-1_ses-test-yaml-1_ecephys.nwb",
-        "sub-yaml-002/sub-yaml-002_ses-test-yaml-2_ecephys.nwb",
-        "sub-YAML-Subject-Name/sub-YAML-Subject-Name_ses-test-yaml-3_ecephys.nwb",
-    ]
-    for asset_path in expected_asset_paths:
-        test_asset = dandiset.get_asset_by_path(path=asset_path)  # Will error if not found
-        test_asset_metadata = test_asset.get_raw_metadata()
-
-        # Past uploads may have created the same apparent file, so look at the modification time to ensure
-        # this test is actually testing the most recent upload
-        date_modified = datetime.fromisoformat(test_asset_metadata["dateModified"])
-        assert datetime.now() - date_modified < timedelta(minutes=10)
